Revert "Add separate fileprefix handlers for vitess crs"
This reverts commit e8275f655a053015a00f2b33bfeea94d61e8ef5d.
