+++ paasta_tools/instance/kubernetes.py
@@ -33,9 +33,7 @@ from paasta_tools import kubernetes_tools
 from paasta_tools import monkrelaycluster_tools
 from paasta_tools import nrtsearchservice_tools
 from paasta_tools import smartstack_tools
-from paasta_tools import vitesscell_tools
 from paasta_tools import vitesscluster_tools
-from paasta_tools import vitesskeyspace_tools
 from paasta_tools.cli.utils import LONG_RUNNING_INSTANCE_TYPE_HANDLERS
 from paasta_tools.instance.hpa_metrics_parser import HPAMetricsDict
 from paasta_tools.instance.hpa_metrics_parser import HPAMetricsParser
@@ -59,8 +57,6 @@ INSTANCE_TYPES_CR = {
     "cassandracluster",
     "kafkacluster",
     "vitesscluster",
-    "vitesscell",
-    "vitesskeyspace",
 }
 INSTANCE_TYPES_K8S = {
     "cassandracluster",
@@ -76,8 +72,6 @@ INSTANCE_TYPE_CR_ID = dict(
     cassandracluster=cassandracluster_tools.cr_id,
     kafkacluster=kafkacluster_tools.cr_id,
     vitesscluster=vitesscluster_tools.cr_id,
-    vitesscell=vitesscell_tools.cr_id,
-    vitesskeyspace=vitesskeyspace_tools.cr_id,
     nrtsearchservice=nrtsearchservice_tools.cr_id,
     nrtsearchserviceeks=nrtsearchservice_tools.cr_id,
     monkrelaycluster=monkrelaycluster_tools.cr_id,
